Creating the Adapter class
Intro:

I will implement the adapter pattern to enable objects with different interfaces to communicate with each other. One example would be the analysis.py and interventions.py. Both files have a separate interface when being called and we can use the adapter pattern whenever we want unrelated classes to work together in a single program. The two interface are not compatible which calls for an adapter design pattern to be implemented.

Solution:

The adapter pattern will let these two classes work together and create compatible interfaces. The adapter class will take no arguments and return nothing in the implementation. The client who wants to call the method signature through analysis.py in interventions.py will now be able to with the adapter wrapper class. Pattern  target