State design pattern #2
The Goal:

This improves the SuD’s design by allowing the object to change its behavior when the internal state of that Object Changes. It will organize the code related to particular states into separate classes. In addition it will simplify the code of the context by eliminating bulky state machine conditionals.